64 bittar och b{ttre bsd fd tjafs..
Rev: lib/simulate.lpc:1.2
Rev: src/BUGS:1.2
Rev: src/Makefile.in:1.3
Rev: src/array.c:1.2
Rev: src/backend.c:1.2
Rev: src/configure.in:1.3
Rev: src/docode.c:1.2
Rev: src/interpret.c:1.2
Rev: src/language.y:1.2
Rev: src/las.h:1.2
Rev: src/lex.c:1.2
Rev: src/list.c:1.2
Rev: src/macros.h:1.2
Rev: src/main.c:1.4
Rev: src/mapping.c:1.2
Rev: src/modules/files/file.c:1.2
Rev: src/modules/files/socket.c:1.2
Rev: src/modules/regexp/regexp.c:1.2
Rev: src/opcodes.c:1.2
Rev: src/port.c:1.3
Rev: src/port.h:1.3
Rev: src/program.c:1.2
Rev: src/rusage.c:1.2
Rev: src/stralloc.c:1.2
Rev: src/stralloc.h:1.2
Rev: src/svalue.c:1.2

+++ src/mapping.c
@@ -272,13 +272,13 @@ void describe_mapping(struct mapping *m,struct processing *p,int indent)
   {
     if(p->pointer_a == (void *)m)
     {
-      sprintf(buf,"@%ld",e);
+      sprintf(buf,"@%ld",(long)e);
       my_strcat(buf);
       return;
     }
   }
   
-  sprintf(buf,"([ /* %ld elements */\n", m->ind->size);
+  sprintf(buf,"([ /* %ld elements */\n", (long) m->ind->size);
   my_strcat(buf);
   for(e=0;e<m->ind->size;e++)
   {
